Create Calculator.py
I used information from: https://www.digitalocean.com/community/tutorials/how-to-make-a-calculator-program-in-python-3 to create this script.

Once this script is run, the user will be able to select an operation of either:
1. Addition
2. Subtraction
3. Multiplication
4. Division
The user then enters either 1, 2, 3, or 4.
For this example, let's say the user entered the number 2 for subtraction.

Once the operation is chosen, the user can enter their first number. For this example, we will use the number 13.

The user can then enter a second number, which will be the number that is subtracted from 13. For this example, we will use the number 5.

The code then inputs the answer to 13 - 5, which is 8.0.

Lastly, the code will ask the user if they would like to perform another calculation, while the user can either answer with "yes" or with "no."

If the user answers with "yes," the entire code will start again and the user will start at the beginning where they can select an operation of 1, 2, 3, or 4.

If the user answers with "no," then the coding will end.